backend/requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.6
google-auth==2.36.0
httpx==0.28.1
pgvector==0.3.6
psycopg[binary]==3.2.3
pydantic-settings==2.7.0
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.19
SQLAlchemy==2.0.36
uvicorn[standard]==0.34.0
pytest==8.3.4
pytest-asyncio==0.25.0

